refactor: divide common DTO into separate packages
